requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
openai-whisper==20231117
paddleocr==2.7.0.3
Pillow==10.2.0
python-dotenv==1.0.1 